refactor: redirect CapabilityOptions to new
