+++ packages/fx-core/src/question/constants.ts
@@ -1,147 +1,11 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { Inputs, OptionItem, Platform } from "@microsoft/teamsfx-api";
+import { Inputs, OptionItem } from "@microsoft/teamsfx-api";
 import { FeatureFlags, featureFlagManager } from "../common/featureFlags";
 import { getLocalizedString } from "../common/localizeUtils";
-import { ProjectTypeOptions } from "./scaffold/vsc/ProjectTypeOptions";
-
-export enum QuestionNames {
-  Scratch = "scratch",
-  SctatchYes = "scratch-yes",
-  AppName = "app-name",
-  Folder = "folder",
-  ProjectPath = "projectPath",
-  ProgrammingLanguage = "programming-language",
-  ProjectType = "project-type",
-  Capabilities = "capabilities",
-  TeamsAppType = "teams-app-type",
-  BotTrigger = "bot-host-type-trigger",
-  Runtime = "runtime",
-  SPFxSolution = "spfx-solution",
-  SPFxInstallPackage = "spfx-install-latest-package",
-  SPFxFramework = "spfx-framework-type",
-  SPFxWebpartName = "spfx-webpart-name",
-  SPFxWebpartDesc = "spfx-webpart-desp",
-  SPFxFolder = "spfx-folder",
-  OfficeAddinFolder = "addin-project-folder",
-  OfficeAddinManifest = "addin-project-manifest",
-  OfficeAddinTemplate = "addin-template-select",
-  OfficeAddinHost = "addin-host",
-  OfficeAddinImport = "addin-import",
-  OfficeAddinFramework = "office-addin-framework-type",
-  Samples = "samples",
-  ReplaceContentUrl = "replaceContentUrl",
-  ReplaceWebsiteUrl = "replaceWebsiteUrl",
-  ReplaceBotIds = "replaceBotIds",
-  SafeProjectName = "safeProjectName",
-  RepalceTabUrl = "tdp-tab-url",
-  ValidateMethod = "validate-method",
-  AppPackagePath = "appPackagePath",
-  FromExistingApi = "from-existing-api", // group name for creating an App from existing api
-  SearchOpenAPISpecQuery = "search-openapi-spec-query",
-  SelectOpenApiSpec = "select-openapi-spec",
-  OpenAPISpecType = "openapi-spec-type",
-  ApiSpecLocation = "openapi-spec-location",
-  ApiOperation = "api-operation",
-  ActionManifestPath = "external-api-plugin-manifest-path", // manifest path for creating project from existing plugin manifest. Use in Kiota integration, etc.
-  MeArchitectureType = "me-architecture",
-  ApiSpecApiKey = "api-key",
-  ApiSpecApiKeyConfirm = "api-key-confirm",
-  ApiAuth = "api-auth",
-  OauthClientSecret = "oauth-client-secret",
-  OauthClientId = "oauth-client-id",
-  OauthConfirm = "oauth-confirm",
-
-  CustomCopilotRag = "custom-copilot-rag",
-  CustomCopilotAssistant = "custom-copilot-agent",
-  LLMService = "llm-service",
-  OpenAIKey = "openai-key",
-  OpenAIEmbeddingModel = "openai-embedding-model",
-  AzureOpenAIKey = "azure-openai-key",
-  AzureOpenAIEndpoint = "azure-openai-endpoint",
-  AzureOpenAIDeploymentName = "azure-openai-deployment-name",
-  AzureOpenAIEmbeddingDeploymentName = "azure-openai-embedding-deployment-name",
-  AzureAISearchApiKey = "azure-ai-search-api-key",
-  AzureAISearchEndpoint = "azure-ai-search-endpoint",
-
-  Features = "features",
-  Env = "env",
-  SourceEnvName = "sourceEnvName",
-  TargetEnvName = "targetEnvName",
-  TargetResourceGroupName = "targetResourceGroupName",
-  NewResourceGroupName = "newResourceGroupName",
-  NewResourceGroupLocation = "newResourceGroupLocation",
-  NewTargetEnvName = "newTargetEnvName",
-  ExistingTabEndpoint = "existing-tab-endpoint",
-  TeamsAppManifestFilePath = "manifest-path",
-  LocalTeamsAppManifestFilePath = "local-manifest-path",
-  AadAppManifestFilePath = "manifest-file-path",
-  TeamsAppPackageFilePath = "app-package-file-path",
-  ConfirmManifest = "confirmManifest",
-  ConfirmLocalManifest = "confirmLocalManifest",
-  ConfirmAadManifest = "confirmAadManifest",
-  OutputZipPathParamName = "output-zip-path",
-  OutputManifestParamName = "output-manifest-path",
-  OutputFolderParamName = "output-folder",
-  M365Host = "m365-host",
-
-  ManifestPath = "manifest-path",
-  ManifestId = "manifest-id",
-  TeamsAppId = "teams-app-id",
-  TitleId = "title-id",
-  UserEmail = "email",
-
-  UninstallMode = "mode",
-  UninstallModeManifestId = "manifest-id",
-  UninstallModeEnv = "env",
-  UninstallModeTitleId = "title-id",
-  UninstallOptions = "options",
-  UninstallOptionM365 = "m365-app",
-  UninstallOptionTDP = "app-registration",
-  UninstallOptionBot = "bot-framework-registration",
-
-  collaborationAppType = "collaborationType",
-  DestinationApiSpecFilePath = "destination-api-spec-location",
-
-  SyncManifest = "sync-manifest",
-  ActionType = "api-plugin-type",
-  WithPlugin = "with-plugin",
-  ImportPlugin = "import-plugin",
-  PluginManifestFilePath = "plugin-manifest-path",
-  PluginOpenApiSpecFilePath = "plugin-opeanapi-spec-path",
-  KnowledgeSource = "knowledge-source",
-  OneDriveSharePointURL = "oneDriveSharePointURL",
-  OneDriveSharePointContent = "oneDriveSharePointContent",
-  WebContent = "web-content",
-  SearchType = "search-type",
-  GCContent = "graph-connector-content",
-  GCList = "graph-connector-list",
-  GCInput = "graph-connector-input",
-  GCName = "graph-connector-name",
-  GCConnectionId = "graph-connector-connection-id",
-  AuthName = "auth-name",
-  TemplateName = "template-name",
-
-  EmbeddedKnowledgeFiles = "embedded-knowledge-files",
-  OAuthAuthorizationUrl = "oauth-authorization-url",
-  OAuthTokenUrl = "oauth-token-url",
-  OAuthRefreshUrl = "oauth-refresh-url",
-  OAuthScope = "oauth-scope",
-  OauthPKCE = "oauth-pkce",
-  ApiKeyIn = "api-key-in",
-  ApiKeyName = "api-key-name",
-
-  TypeSpecProjectType = "type-spec-project-type",
-  DeclarativeAgentManifestPath = "declarative-agent-manifest-path",
-  SensitivityLabel = "sensitivity-label",
-
-  ShareOption = "option",
-  ShareOptionShareApp = "share-app",
-  ShareOptionShareToUser = "share-with-users",
-  ShareToUsers = "users",
-  RemoveUsers = "users",
-}
+import { QuestionNames } from "./questionNames";
+export { QuestionNames } from "./questionNames";
 
 export const AppNamePattern =
   '^(?=(.*[\\da-zA-Z]){2})[a-zA-Z][^"<>:\\?/*&|\u0000-\u001F]*[^"\\s.<>:\\?/*&|\u0000-\u001F]$';
@@ -169,33 +33,6 @@ export const capabilitiesHavePythonOption = [
   "custom-copilot-rag-customApi",
 ];
 
-export class RuntimeOptions {
-  static NodeJS(): OptionItem {
-    return {
-      id: "node",
-      label: "Node.js",
-      detail: getLocalizedString("core.RuntimeOptionNodeJS.detail"),
-    };
-  }
-  static DotNet(): OptionItem {
-    return {
-      id: "dotnet",
-      label: ".NET Core",
-      detail: getLocalizedString("core.RuntimeOptionDotNet.detail"),
-    };
-  }
-}
-
-export function getRuntime(inputs: Inputs): string {
-  let runtime = RuntimeOptions.NodeJS().id;
-  if (inputs?.platform === Platform.VS) {
-    runtime = RuntimeOptions.DotNet().id;
-  } else if (featureFlagManager.getBooleanValue(FeatureFlags.CLIDotNet)) {
-    runtime = inputs[QuestionNames.Runtime] || runtime;
-  }
-  return runtime;
-}
-
 export class ScratchOptions {
   static yes(): OptionItem {
     return {
@@ -213,489 +50,6 @@ export class ScratchOptions {
   }
   static all(): OptionItem[] {
     return [ScratchOptions.yes(), ScratchOptions.no()];
-  }
-}
-
-export class CapabilityOptions {
-  // empty
-  static empty(): OptionItem {
-    return {
-      id: "empty",
-      label: "Empty",
-    };
-  }
-
-  // bot
-  static basicBot(): OptionItem {
-    return {
-      id: "bot",
-      label: `${getLocalizedString("core.BotNewUIOption.label")}`,
-      detail: getLocalizedString("core.BotNewUIOption.detail"),
-    };
-  }
-  static notificationBot(): OptionItem {
-    return {
-      // For default option, id and cliName must be the same
-      id: "notification",
-      label: `${getLocalizedString("core.NotificationOption.label")}`,
-      detail: getLocalizedString("core.NotificationOption.detail"),
-      data: "https://aka.ms/teamsfx-send-notification",
-      buttons: [
-        {
-          iconPath: "file-symlink-file",
-          tooltip: getLocalizedString("core.option.github"),
-          command: "fx-extension.openTutorial",
-        },
-      ],
-    };
-  }
-
-  static commandBot(): OptionItem {
-    return {
-      // id must match cli `yargsHelp`
-      id: "command-bot",
-      label: `${getLocalizedString("core.CommandAndResponseOption.label")}`,
-      detail: getLocalizedString("core.CommandAndResponseOption.detail"),
-      data: "https://aka.ms/teamsfx-create-command",
-      buttons: [
-        {
-          iconPath: "file-symlink-file",
-          tooltip: getLocalizedString("core.option.github"),
-          command: "fx-extension.openTutorial",
-        },
-      ],
-    };
-  }
-
-  static workflowBot(inputs?: Inputs): OptionItem {
-    const item: OptionItem = {
-      // id must match cli `yargsHelp`
-      id: "workflow-bot",
-      label: `${getLocalizedString("core.WorkflowOption.label")}`,
-      detail: getLocalizedString("core.WorkflowOption.detail"),
-      data: "https://aka.ms/teamsfx-create-workflow",
-      buttons: [
-        {
-          iconPath: "file-symlink-file",
-          tooltip: getLocalizedString("core.option.github"),
-          command: "fx-extension.openTutorial",
-        },
-      ],
-    };
-    if (inputs?.inProductDoc) {
-      item.data = "cardActionResponse";
-      item.buttons = [
-        {
-          iconPath: "file-code",
-          tooltip: getLocalizedString("core.option.inProduct"),
-          command: "fx-extension.openTutorial",
-        },
-      ];
-    }
-    return item;
-  }
-
-  //tab
-
-  static nonSsoTab(): OptionItem {
-    return {
-      id: "tab-non-sso",
-      label: `${getLocalizedString("core.TabNonSso.label")}`,
-      detail: getLocalizedString("core.TabNonSso.detail"),
-      description: getLocalizedString(
-        "core.createProjectQuestion.option.description.worksInOutlookM365"
-      ),
-    };
-  }
-
-  static tab(): OptionItem {
-    return {
-      id: "tab",
-      label: getLocalizedString("core.TabOption.label"),
-      description: getLocalizedString("core.TabOption.description"),
-      detail: getLocalizedString("core.TabOption.detail"),
-    };
-  }
-
-  static m365SsoLaunchPage(): OptionItem {
-    return {
-      id: "sso-launch-page",
-      label: `${getLocalizedString("core.M365SsoLaunchPageOptionItem.label")}`,
-      detail: getLocalizedString("core.M365SsoLaunchPageOptionItem.detail"),
-      description: getLocalizedString(
-        "core.createProjectQuestion.option.description.worksInOutlookM365"
-      ),
-    };
-  }
-
-  static dashboardTab(): OptionItem {
-    return {
-      id: "dashboard-tab",
-      label: `${getLocalizedString("core.DashboardOption.label")}`,
-      detail: getLocalizedString("core.DashboardOption.detail"),
-      description: getLocalizedString(
-        "core.createProjectQuestion.option.description.worksInOutlookM365"
-      ),
-      data: "https://aka.ms/teamsfx-dashboard-app",
-      buttons: [
-        {
-          iconPath: "file-symlink-file",
-          tooltip: getLocalizedString("core.option.github"),
-          command: "fx-extension.openTutorial",
-        },
-      ],
-    };
-  }
-
-  static SPFxTab(): OptionItem {
-    return {
-      id: "tab-spfx",
-      label: getLocalizedString("core.TabSPFxOption.labelNew"),
-      description: getLocalizedString(
-        "core.createProjectQuestion.option.description.worksInOutlookM365"
-      ),
-      detail: getLocalizedString("core.TabSPFxOption.detailNew"),
-    };
-  }
-
-  //message extension
-  static linkUnfurling(): OptionItem {
-    return {
-      id: "link-unfurling",
-      label: `${getLocalizedString("core.LinkUnfurlingOption.label")}`,
-      detail: getLocalizedString("core.LinkUnfurlingOption.detail"),
-      description: getLocalizedString(
-        "core.createProjectQuestion.option.description.worksInOutlook"
-      ),
-    };
-  }
-
-  static m365SearchMe(): OptionItem {
-    return {
-      id: "search-app",
-      label: `${getLocalizedString("core.M365SearchAppOptionItem.label")}`,
-      detail: getLocalizedString("core.M365SearchAppOptionItem.copilot.detail"),
-    };
-  }
-
-  static SearchMe(): OptionItem {
-    return {
-      id: "search-message-extension",
-      label: `${getLocalizedString("core.M365SearchAppOptionItem.label")}`,
-      detail: getLocalizedString("core.SearchAppOptionItem.detail"),
-    };
-  }
-
-  static collectFormMe(): OptionItem {
-    return {
-      id: "collect-form-message-extension",
-      label: `${getLocalizedString("core.MessageExtensionOption.labelNew")}`,
-      detail: getLocalizedString("core.MessageExtensionOption.detail"),
-    };
-  }
-  static me(): OptionItem {
-    return {
-      id: "message-extension",
-      label: getLocalizedString("core.MessageExtensionOption.label"),
-      description: getLocalizedString("core.MessageExtensionOption.description"),
-      detail: getLocalizedString("core.MessageExtensionOption.detail"),
-    };
-  }
-  static bots(inputs?: Inputs): OptionItem[] {
-    if (inputs && getRuntime(inputs) === RuntimeOptions.DotNet().id) {
-      return [
-        CapabilityOptions.basicBot(),
-        CapabilityOptions.aiBot(),
-        CapabilityOptions.aiAssistantBot(),
-        CapabilityOptions.notificationBot(),
-        CapabilityOptions.commandBot(),
-        CapabilityOptions.workflowBot(inputs),
-      ];
-    }
-    return [
-      CapabilityOptions.basicBot(),
-      CapabilityOptions.notificationBot(),
-      CapabilityOptions.commandBot(),
-      CapabilityOptions.workflowBot(inputs),
-    ];
-  }
-
-  static tabs(): OptionItem[] {
-    return [
-      CapabilityOptions.nonSsoTab(),
-      CapabilityOptions.m365SsoLaunchPage(),
-      CapabilityOptions.dashboardTab(),
-      CapabilityOptions.SPFxTab(),
-    ];
-  }
-
-  static dotnetCaps(inputs?: Inputs): OptionItem[] {
-    const capabilities = [
-      CapabilityOptions.empty(),
-      ...CapabilityOptions.agents(),
-      ...CapabilityOptions.customCopilots(inputs),
-      ...CapabilityOptions.bots(inputs),
-      CapabilityOptions.nonSsoTab(),
-      CapabilityOptions.tab(),
-      ...CapabilityOptions.collectMECaps(),
-    ];
-    if (featureFlagManager.getBooleanValue(FeatureFlags.TdpTemplateCliTest)) {
-      capabilities.push(CapabilityOptions.me());
-    }
-
-    return capabilities;
-  }
-
-  /**
-   * Collect all capabilities for message extension, including dotnet and nodejs.
-   * @returns OptionItem[] capability list
-   */
-  static collectMECaps(): OptionItem[] {
-    return [
-      CapabilityOptions.m365SearchMe(),
-      CapabilityOptions.collectFormMe(),
-      CapabilityOptions.SearchMe(),
-      CapabilityOptions.linkUnfurling(),
-    ];
-  }
-
-  static mes(inputs?: Inputs): OptionItem[] {
-    return inputs !== undefined && getRuntime(inputs) === RuntimeOptions.DotNet().id
-      ? [
-          CapabilityOptions.SearchMe(),
-          CapabilityOptions.collectFormMe(),
-          CapabilityOptions.linkUnfurling(),
-        ]
-      : [
-          CapabilityOptions.m365SearchMe(),
-          CapabilityOptions.collectFormMe(),
-          CapabilityOptions.linkUnfurling(),
-        ];
-  }
-
-  static officeAddinCapabilities(projectType: string): OptionItem[] {
-    const items: OptionItem[] = [CapabilityOptions.officeAddinTaskpane()];
-    const isOutlookAddin = projectType === ProjectTypeOptions.outlookAddin().id;
-    const isMetaOSAddin = projectType === ProjectTypeOptions.officeMetaOS().id;
-    if (isOutlookAddin) {
-      items.push(CapabilityOptions.outlookAddinImport());
-    } else if (isMetaOSAddin) {
-      items.push(CapabilityOptions.officeAddinImport());
-    }
-    return items;
-  }
-
-  static agents(): OptionItem[] {
-    return [CapabilityOptions.declarativeAgent()];
-  }
-
-  static customCopilots(inputs?: Inputs): OptionItem[] {
-    return inputs !== undefined && getRuntime(inputs) === RuntimeOptions.DotNet().id
-      ? [
-          CapabilityOptions.customCopilotBasic(),
-          CapabilityOptions.customCopilotRag(),
-          CapabilityOptions.customCopilotAssistant(),
-          // customCopilotWeather is currently only for dotnet
-          CapabilityOptions.customCopilotWeather(),
-        ]
-      : [
-          CapabilityOptions.customCopilotBasic(),
-          CapabilityOptions.customCopilotRag(),
-          CapabilityOptions.customCopilotAssistant(),
-        ];
-  }
-
-  static tdpIntegrationCapabilities(): OptionItem[] {
-    // templates that are used by TDP integration only
-    return [
-      CapabilityOptions.me(),
-      CapabilityOptions.botAndMe(),
-      CapabilityOptions.nonSsoTabAndBot(),
-    ];
-  }
-
-  /**
-   * static capability list, which does not depend on any feature flags
-   */
-  static staticAll(inputs?: Inputs): OptionItem[] {
-    const capabilityOptions = [
-      CapabilityOptions.empty(),
-      ...CapabilityOptions.bots(inputs),
-      ...CapabilityOptions.tabs(),
-      ...CapabilityOptions.collectMECaps(),
-      ...CapabilityOptions.agents(),
-      ...CapabilityOptions.customCopilots(),
-      ...CapabilityOptions.tdpIntegrationCapabilities(),
-    ];
-    return capabilityOptions;
-  }
-
-  /**
-   * dynamic capability list, which depends on feature flags
-   */
-  static all(inputs?: Inputs): OptionItem[] {
-    if (inputs && getRuntime(inputs) === RuntimeOptions.DotNet().id) {
-      return CapabilityOptions.dotnetCaps(inputs);
-    }
-    const capabilityOptions = [
-      ...CapabilityOptions.bots(inputs),
-      ...CapabilityOptions.tabs(),
-      ...CapabilityOptions.collectMECaps(),
-    ];
-    capabilityOptions.push(...CapabilityOptions.agents());
-    capabilityOptions.push(...CapabilityOptions.customCopilots());
-    if (featureFlagManager.getBooleanValue(FeatureFlags.TdpTemplateCliTest)) {
-      // test templates that are used by TDP integration only
-      capabilityOptions.push(...CapabilityOptions.tdpIntegrationCapabilities());
-    }
-    capabilityOptions.push(...CapabilityOptions.officeAddinCapabilities(inputs?.projectType));
-    return capabilityOptions;
-  }
-
-  static officeAddinTaskpane(): OptionItem {
-    return {
-      id: "office-addin-taskpane",
-      label: getLocalizedString("core.newTaskpaneAddin.label"),
-      detail: getLocalizedString("core.newTaskpaneAddin.detail"),
-    };
-  }
-
-  static outlookAddinImport(): OptionItem {
-    return {
-      id: "outlook-addin-import",
-      label: getLocalizedString("core.importAddin.label"),
-      detail: getLocalizedString("core.importAddin.detail"),
-    };
-  }
-
-  static officeAddinImport(): OptionItem {
-    return {
-      id: "office-addin-import",
-      label: getLocalizedString("core.importOfficeAddin.label"),
-      detail: getLocalizedString("core.importAddin.detail"),
-      description: getLocalizedString(
-        "core.createProjectQuestion.option.description.previewOnWindow"
-      ),
-    };
-  }
-
-  static officeContentAddin(): OptionItem {
-    return {
-      id: "office-content-addin",
-      label: getLocalizedString("core.officeContentAddin.label"),
-      detail: getLocalizedString("core.officeContentAddin.detail"),
-    };
-  }
-
-  static nonSsoTabAndBot(): OptionItem {
-    return {
-      id: "TabNonSsoAndBot",
-      label: "", // No need to set display name as this option won't be shown in UI
-    };
-  }
-
-  static botAndMe(): OptionItem {
-    return {
-      id: "BotAndMessageExtension",
-      label: "", // No need to set display name as this option won't be shown in UI
-    };
-  }
-
-  // copilot extension - api plugin
-  static apiPlugin(): OptionItem {
-    return {
-      id: "api-plugin",
-      label: getLocalizedString("core.createProjectQuestion.projectType.copilotPlugin.label"),
-      detail: getLocalizedString("core.createProjectQuestion.projectType.copilotPlugin.detail"),
-    };
-  }
-
-  // copilot extension - declarative copilot
-  static declarativeAgent(): OptionItem {
-    return {
-      id: "declarative-agent",
-      label: getLocalizedString("core.createProjectQuestion.projectType.declarativeAgent.label"),
-      detail: getLocalizedString("core.createProjectQuestion.projectType.declarativeAgent.detail"),
-    };
-  }
-
-  static aiBot(): OptionItem {
-    return {
-      id: "ai-bot",
-      label: getLocalizedString("core.aiBotOption.label"),
-      detail: getLocalizedString("core.aiBotOption.detail"),
-    };
-  }
-
-  static aiAssistantBot(): OptionItem {
-    return {
-      id: "ai-assistant-bot",
-      label: getLocalizedString("core.aiAssistantBotOption.label"),
-      detail: getLocalizedString("core.aiAssistantBotOption.detail"),
-      description: getLocalizedString("core.createProjectQuestion.option.description.preview"),
-    };
-  }
-
-  // custom copilot
-  static customCopilotBasic(): OptionItem {
-    const description = featureFlagManager.getBooleanValue(FeatureFlags.CEAEnabled)
-      ? getLocalizedString("core.createProjectQuestion.capability.customEngineAgent.description")
-      : undefined;
-    return {
-      id: "custom-copilot-basic",
-      label: getLocalizedString(
-        "core.createProjectQuestion.capability.customCopilotBasicOption.label"
-      ),
-      detail: getLocalizedString(
-        "core.createProjectQuestion.capability.customCopilotBasicOption.detail"
-      ),
-      description: description,
-    };
-  }
-
-  static customCopilotRag(): OptionItem {
-    const description = featureFlagManager.getBooleanValue(FeatureFlags.CEAEnabled)
-      ? getLocalizedString("core.createProjectQuestion.capability.customEngineAgent.description")
-      : undefined;
-    return {
-      id: "custom-copilot-rag",
-      label: getLocalizedString(
-        "core.createProjectQuestion.capability.customCopilotRagOption.label"
-      ),
-      detail: getLocalizedString(
-        "core.createProjectQuestion.capability.customCopilotRagOption.detail"
-      ),
-      description: description,
-    };
-  }
-
-  static customCopilotAssistant(): OptionItem {
-    const description = featureFlagManager.getBooleanValue(FeatureFlags.CEAEnabled)
-      ? getLocalizedString("core.createProjectQuestion.capability.customEngineAgent.description")
-      : undefined;
-    return {
-      id: "custom-copilot-agent",
-      label: getLocalizedString(
-        "core.createProjectQuestion.capability.customCopilotAssistantOption.label"
-      ),
-      detail: getLocalizedString(
-        "core.createProjectQuestion.capability.customCopilotAssistantOption.detail"
-      ),
-      description: description,
-    };
-  }
-
-  // Currently only for dotnet
-  static customCopilotWeather(): OptionItem {
-    return {
-      id: "custom-copilot-weather-agent",
-      label: getLocalizedString(
-        "core.createProjectQuestion.capability.customCopilotWeatherOption.label"
-      ),
-      detail: getLocalizedString(
-        "core.createProjectQuestion.capability.customCopilotWeatherOption.detail"
-      ),
-    };
   }
 }
