+++ packages/fx-core/src/question/questionNames.ts
@@ -1,0 +1,138 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+export enum QuestionNames {
+  Scratch = "scratch",
+  SctatchYes = "scratch-yes",
+  AppName = "app-name",
+  Folder = "folder",
+  ProjectPath = "projectPath",
+  ProgrammingLanguage = "programming-language",
+  ProjectType = "project-type",
+  Capabilities = "capabilities",
+  TeamsAppType = "teams-app-type",
+  BotTrigger = "bot-host-type-trigger",
+  Runtime = "runtime",
+  SPFxSolution = "spfx-solution",
+  SPFxInstallPackage = "spfx-install-latest-package",
+  SPFxFramework = "spfx-framework-type",
+  SPFxWebpartName = "spfx-webpart-name",
+  SPFxWebpartDesc = "spfx-webpart-desp",
+  SPFxFolder = "spfx-folder",
+  OfficeAddinFolder = "addin-project-folder",
+  OfficeAddinManifest = "addin-project-manifest",
+  OfficeAddinTemplate = "addin-template-select",
+  OfficeAddinHost = "addin-host",
+  OfficeAddinImport = "addin-import",
+  OfficeAddinFramework = "office-addin-framework-type",
+  Samples = "samples",
+  ReplaceContentUrl = "replaceContentUrl",
+  ReplaceWebsiteUrl = "replaceWebsiteUrl",
+  ReplaceBotIds = "replaceBotIds",
+  SafeProjectName = "safeProjectName",
+  RepalceTabUrl = "tdp-tab-url",
+  ValidateMethod = "validate-method",
+  AppPackagePath = "appPackagePath",
+  FromExistingApi = "from-existing-api", // group name for creating an App from existing api
+  SearchOpenAPISpecQuery = "search-openapi-spec-query",
+  SelectOpenApiSpec = "select-openapi-spec",
+  OpenAPISpecType = "openapi-spec-type",
+  ApiSpecLocation = "openapi-spec-location",
+  ApiOperation = "api-operation",
+  ActionManifestPath = "external-api-plugin-manifest-path", // manifest path for creating project from existing plugin manifest. Use in Kiota integration, etc.
+  MeArchitectureType = "me-architecture",
+  ApiSpecApiKey = "api-key",
+  ApiSpecApiKeyConfirm = "api-key-confirm",
+  ApiAuth = "api-auth",
+  OauthClientSecret = "oauth-client-secret",
+  OauthClientId = "oauth-client-id",
+  OauthConfirm = "oauth-confirm",
+
+  CustomCopilotRag = "custom-copilot-rag",
+  CustomCopilotAssistant = "custom-copilot-agent",
+  LLMService = "llm-service",
+  OpenAIKey = "openai-key",
+  OpenAIEmbeddingModel = "openai-embedding-model",
+  AzureOpenAIKey = "azure-openai-key",
+  AzureOpenAIEndpoint = "azure-openai-endpoint",
+  AzureOpenAIDeploymentName = "azure-openai-deployment-name",
+  AzureOpenAIEmbeddingDeploymentName = "azure-openai-embedding-deployment-name",
+  AzureAISearchApiKey = "azure-ai-search-api-key",
+  AzureAISearchEndpoint = "azure-ai-search-endpoint",
+
+  Features = "features",
+  Env = "env",
+  SourceEnvName = "sourceEnvName",
+  TargetEnvName = "targetEnvName",
+  TargetResourceGroupName = "targetResourceGroupName",
+  NewResourceGroupName = "newResourceGroupName",
+  NewResourceGroupLocation = "newResourceGroupLocation",
+  NewTargetEnvName = "newTargetEnvName",
+  ExistingTabEndpoint = "existing-tab-endpoint",
+  TeamsAppManifestFilePath = "manifest-path",
+  LocalTeamsAppManifestFilePath = "local-manifest-path",
+  AadAppManifestFilePath = "manifest-file-path",
+  TeamsAppPackageFilePath = "app-package-file-path",
+  ConfirmManifest = "confirmManifest",
+  ConfirmLocalManifest = "confirmLocalManifest",
+  ConfirmAadManifest = "confirmAadManifest",
+  OutputZipPathParamName = "output-zip-path",
+  OutputManifestParamName = "output-manifest-path",
+  OutputFolderParamName = "output-folder",
+  M365Host = "m365-host",
+
+  ManifestPath = "manifest-path",
+  ManifestId = "manifest-id",
+  TeamsAppId = "teams-app-id",
+  TitleId = "title-id",
+  UserEmail = "email",
+
+  UninstallMode = "mode",
+  UninstallModeManifestId = "manifest-id",
+  UninstallModeEnv = "env",
+  UninstallModeTitleId = "title-id",
+  UninstallOptions = "options",
+  UninstallOptionM365 = "m365-app",
+  UninstallOptionTDP = "app-registration",
+  UninstallOptionBot = "bot-framework-registration",
+
+  collaborationAppType = "collaborationType",
+  DestinationApiSpecFilePath = "destination-api-spec-location",
+
+  SyncManifest = "sync-manifest",
+  ActionType = "api-plugin-type",
+  WithPlugin = "with-plugin",
+  ImportPlugin = "import-plugin",
+  PluginManifestFilePath = "plugin-manifest-path",
+  PluginOpenApiSpecFilePath = "plugin-opeanapi-spec-path",
+  KnowledgeSource = "knowledge-source",
+  OneDriveSharePointURL = "oneDriveSharePointURL",
+  OneDriveSharePointContent = "oneDriveSharePointContent",
+  WebContent = "web-content",
+  SearchType = "search-type",
+  GCContent = "graph-connector-content",
+  GCList = "graph-connector-list",
+  GCInput = "graph-connector-input",
+  GCName = "graph-connector-name",
+  GCConnectionId = "graph-connector-connection-id",
+  AuthName = "auth-name",
+  TemplateName = "template-name",
+
+  EmbeddedKnowledgeFiles = "embedded-knowledge-files",
+  OAuthAuthorizationUrl = "oauth-authorization-url",
+  OAuthTokenUrl = "oauth-token-url",
+  OAuthRefreshUrl = "oauth-refresh-url",
+  OAuthScope = "oauth-scope",
+  OauthPKCE = "oauth-pkce",
+  ApiKeyIn = "api-key-in",
+  ApiKeyName = "api-key-name",
+
+  TypeSpecProjectType = "type-spec-project-type",
+  DeclarativeAgentManifestPath = "declarative-agent-manifest-path",
+  SensitivityLabel = "sensitivity-label",
+
+  ShareOption = "option",
+  ShareOptionShareApp = "share-app",
+  ShareOptionShareToUser = "share-with-users",
+  ShareToUsers = "users",
+  RemoveUsers = "users",
+}
