+++ packages/fx-core/src/question/scaffold/vsc/createRootNode.ts
@@ -5,8 +5,9 @@ import { Inputs, IQTreeNode, OptionItem, Platform } from "@microsoft/teamsfx-api
 import { featureFlagManager, FeatureFlags } from "../../../common/featureFlags";
 import { getLocalizedString } from "../../../common/localizeUtils";
 import { getAllTemplatesOnPlatform } from "../../../component/generator/templates/metadata";
-import { ProgrammingLanguage, QuestionNames } from "../../constants";
+import { ProgrammingLanguage } from "../../constants";
 import { appNameQuestion, folderQuestion } from "../../create";
+import { QuestionNames } from "../../questionNames";
 import {
   ActionStartOptions,
   BotCapabilityOptions,
